merapihkan code

+++ day16/index.js
@@ -1,3 +1,4 @@
+// App Initialization
 const express = require('express')
 const path = require('path')
 const app = express()
@@ -12,10 +13,14 @@ const session = require('express-session')
 const flash = require('express-flash')
 const blogModel = require('./src/models').blog
 const upload = require('./src/middlewares/uploadFile')
+const functions = require('./src/utilities/functions')
+const { formatDate, getDefaultImage, getSelectedIcons, dayDifference, chooseDuration } = require('./src/utilities/functions')
 
+// Set view engine and views directory
 app.set("view engine", "hbs")
 app.set("views", path.join(__dirname, 'src/views'))
 
+// Static file and body parser middleware
 app.use("/assets", express.static('src/assets'))
 app.use("/uploads", express.static('src/uploads'))
 app.use(express.urlencoded({ extended: false }))
@@ -30,8 +35,10 @@ app.use(session({
     }
 }))
 
-handlebars.registerHelper('eq', handlebarsEqual);
+// Handlebars helper registration
+handlebars.registerHelper('eq', handlebarsEqual)
 
+// Routes
 app.get('/', home)
 app.get('/contact', checkAuth, contact)
 app.post('/delete-card/:id', checkAuth, deleteCard)
@@ -53,6 +60,7 @@ app.post('/login', login)
 
 const data = []
 
+// Function to render home page
 async function home(req, res) {
     const query = `SELECT projects.id, projects.name, projects.start_date,
     projects.end_date, projects.description, projects.technologies,
@@ -85,6 +93,7 @@ async function home(req, res) {
     res.render('index', { data: projectsWithInfo, user: req.session.user, isLogin: isLogin, user: user })
 }
 
+// Function to render contact page
 function contact(req, res) {
     const isLogin = req.session.isLogin
     const user = req.session.user
@@ -92,6 +101,7 @@ function contact(req, res) {
     res.render('contact', {isLogin: isLogin, user: user})
 }
 
+// Function to render add project view
 function addProjectView(req, res) {
     const isLogin = req.session.isLogin
     const user = req.session.user
@@ -99,6 +109,7 @@ function addProjectView(req, res) {
     res.render('add-project', {isLogin: isLogin, user: user})
 }
 
+// Function to handle project addition
 async function addProject(req, res) {
     const { name, start, end, description, id } = req.body
     const checkboxes = ['nodeJs', 'nextJs', 'reactJs', 'typeScript'].filter(checkbox => req.body[checkbox])
@@ -106,19 +117,14 @@ async function addProject(req, res) {
     const image = req.file.filename
     const author_id = req.session.user.id
     
-        // Hitung selisih dalam hari
+         // Calculate the difference in days
         const days = dayDifference(new Date(start), new Date(end))
 
-        // Tentukan unit durasi berdasarkan selisih waktu
+         // Determine the duration unit based on the time difference
         const { duration, unit } = chooseDuration(days)
 
-        // Ambil icon yang dipilih
-        const icons = {
-            nodeJs: checkboxes.includes('nodeJs') ? "/assets/img/nodejs.png" : '',
-            nextJs: checkboxes.includes('nextJs') ? "/assets/img/nextjs.png" : '',
-            reactJs: checkboxes.includes('reactJs') ? "/assets/img/react.png"  : '',
-            typeScript: checkboxes.includes('typeScript') ? "/assets/img/typescirpt.png" : ''
-        }
+        // Get the selected icons
+        const icons = getSelectedIcons(checkboxes)
 
         const query = `INSERT INTO projects(name, start_date, end_date, description, technologies, image,author_id) VALUES('${name}', '${start}', '${end}', '${description}', ARRAY['${checkboxes.join("','")}'], '${image}', '${author_id}')`
         const obj = await sequelize.query(query, { type: QueryTypes.INSERT})
@@ -126,6 +132,7 @@ async function addProject(req, res) {
         res.redirect('/')
 }
 
+// Function to render update project view
 async function updateProjectView(req, res) {
     const {id} = req.params
 
@@ -138,53 +145,77 @@ async function updateProjectView(req, res) {
     res.render('update-project', { data: obj[0], isLogin: isLogin, user: user })
 }
 
+// Function to handle project update
 async function updateProject(req, res) {
-    const { name, start, end, description, id } = req.body
-    const checkboxes = ['nodeJs', 'nextJs', 'reactJs', 'typeScript'].filter(checkbox => req.body[checkbox])
+    try {
+        const { name, start, end, description, id } = req.body;
+        const checkboxes = ['nodeJs', 'nextJs', 'reactJs', 'typeScript'].filter(checkbox => req.body[checkbox]);
 
+        // Calculate the difference in days
+        const days = dayDifference(new Date(start), new Date(end));
 
-        // Hitung selisih dalam hari
-        const days = dayDifference(new Date(start), new Date(end))
+        // Determine the duration unit based on the time difference
+        const { duration, unit } = chooseDuration(days);
 
-        // Tentukan unit durasi berdasarkan selisih waktu
-        const { duration, unit } = chooseDuration(days)
+        // Get the selected icons
+        const icons = getSelectedIcons(checkboxes)
 
-        // Ambil icon yang dipilih
-        const icons = {
-            nodeJs: checkboxes.includes('nodeJs') ? "/assets/img/nodejs.png" : '',
-            nextJs: checkboxes.includes('nextJs') ? "/assets/img/nextjs.png" : '',
-            reactJs: checkboxes.includes('reactJs') ? "/assets/img/react.png"  : '',
-            typeScript: checkboxes.includes('typeScript') ? "/assets/img/typescirpt.png" : ''
+        let image = "";
+        if (req.file) {
+            image = req.file.filename;
         }
 
-        let image = ""
-    if (req.file) {
-        image = req.file.filename
+        if (!image) {
+            image = await getDefaultImage(id)
+        }
+
+        // Check project ownership before displaying the update form
+        const projectOwnershipCheckQuery = `SELECT * FROM projects WHERE id=${id} AND author_id=${req.session.user.id}`;
+        const projectOwnershipCheck = await sequelize.query(projectOwnershipCheckQuery, { type: QueryTypes.SELECT });
+
+        if (!projectOwnershipCheck.length) {
+            // Project is not owned by the currently logged-in user
+            req.flash('danger', 'You do not have permission to edit this project.');
+            return res.redirect('/');
+        }
+
+        // Render the update form if granted permission
+        const isLogin = req.session.isLogin;
+        const user = req.session.user;
+
+        const updateQuery = `UPDATE projects SET name='${name}', start_date='${start}', end_date='${end}', description='${description}', technologies=ARRAY['${checkboxes.join("','")}'], image='${image}'  WHERE id=${id}`;
+        const updateResult = await sequelize.query(updateQuery, { type: QueryTypes.UPDATE });
+
+        res.redirect('/');
+    } catch (error) {
+        console.error(error);
+        res.status(500).send('Internal Server Error');
     }
-
-    if (!image) {
-        const query = `SELECT projects.id, projects.title, projects.content, projects.image, 
-        users.name AS author, projects."createdAt", projects."updatedAt" FROM projects LEFT JOIN users ON
-        projects."authorId" = users.id WHERE projects.id=${id}`
-        const obj = await sequelize.query(query, { type: QueryTypes.SELECT })
-        image = obj[0].image
-    }
-
-        const query = `UPDATE projects SET name='${name}', start_date='${start}', end_date='${end}', description='${description}', technologies=ARRAY['${checkboxes.join("','")}'], image='${image}'  WHERE id=${id}`
-        const obj = await sequelize.query(query, { type: QueryTypes.UPDATE })
-
-        res.redirect('/')
 }
 
+// Function to handle project deletion
 async function deleteCard(req, res) {
-    const {id} = req.params
- 
-    const query = `DELETE FROM projects WHERE id=${id}`
-    const obj = await sequelize.query(query, { type: QueryTypes.DELETE })
+    const { id } = req.params;
 
-    res.redirect('/')
+    // Ensure the project is owned by the currently logged-in user
+    const projectOwnershipCheckQuery = `SELECT * FROM projects WHERE id=${id} AND author_id=${req.session.user.id}`;
+    const projectOwnershipCheck = await sequelize.query(projectOwnershipCheckQuery, { type: QueryTypes.SELECT });
+
+    if (!projectOwnershipCheck.length) {
+        // Project is not owned by the currently logged-in user
+        req.flash('danger', 'You do not have permission to delete this project.');
+        return res.redirect('/');
+    }
+
+    // Delete the project if granted permission
+    const query = `DELETE FROM projects WHERE id=${id}`;
+    const obj = await sequelize.query(query, { type: QueryTypes.DELETE });
+
+    req.flash('success', 'Project deleted successfully.');
+    res.redirect('/');
 }
 
+// Function to render project detail
 async function detail(req, res) {
     const { id } = req.params;
 
@@ -196,13 +227,13 @@ async function detail(req, res) {
     console.log('obj:', obj)
 
     if (obj[0] && obj[0].start_date && obj[0].end_date) {
-        // Ubah string menjadi objek Date
+        // Convert strings to Date objects
         const startDate = new Date(obj[0].start_date);
         const endDate = new Date(obj[0].end_date);
 
-        // Format tanggal ke YYYY-MM-DD
-        const formattedStartDate = startDate.toISOString().split('T')[0];
-        const formattedEndDate = endDate.toISOString().split('T')[0];
+        // Format the date to YYYY-MM-DD
+        const formattedStartDate = formatDate(startDate);
+        const formattedEndDate = formatDate(endDate);
 
         const days = dayDifference(startDate, endDate);
         const { duration, unit } = chooseDuration(days);
@@ -212,7 +243,7 @@ async function detail(req, res) {
             nextJs: obj[0].technologies.includes('nextJs') ? "/assets/img/nextjs.png" : '',
             reactJs: obj[0].technologies.includes('reactJs') ? "/assets/img/react.png" : '',
             typeScript: obj[0].technologies.includes('typeScript') ? "/assets/img/typescirpt.png" : ''
-        };
+        }
 
         const isLogin = req.session.isLogin;
         const user = req.session.user;
@@ -234,6 +265,7 @@ async function detail(req, res) {
     }
 }
 
+// Function to render testimonials page
 function testimonials(req, res) {
     const isLogin = req.session.isLogin
     const user = req.session.user
@@ -241,6 +273,7 @@ function testimonials(req, res) {
     res.render('testimonials', {isLogin: isLogin, user: user})
 }
 
+// Function to handle user registration
 async function register(req, res) {
     const {name, email, password} = req.body
 
@@ -265,10 +298,12 @@ async function register(req, res) {
     })
 }
 
+// Function to render registration page
 function registerView(req, res) {
     res.render('register')
 }
 
+// Function to handle user login
 async function login(req, res) {
     const { email, password } = req.body
     const query = `SELECT * FROM users WHERE email='${email}'`
@@ -306,23 +341,24 @@ async function login(req, res) {
     })
 }
 
+// Function to render login page
 function loginView(req, res) {
     res.render('login')
 }
 
-// Middleware untuk memeriksa apakah pengguna sudah login atau belum
+// Middleware to check user authentication
 function checkAuth(req, res, next) {
     if (req.session.isLogin) {
-        // Jika pengguna sudah login, lanjutkan ke route berikutnya
+        // If the user is already logged in, proceed to the next route
         next()
     } else {
-        // Jika pengguna belum login, redirect ke halaman login
+        // If the user is not logged in, redirect to the login page
         req.flash('danger', 'You must be logged in to access this page.')
         res.redirect('/login')
     }
 }
 
-// Ketika tombol logout diklik
+// When the logout button is clicked
 app.post('/logout', (req, res) => {
     req.session.destroy(err => {
         if (err) {
@@ -332,27 +368,7 @@ app.post('/logout', (req, res) => {
     })
 })
 
-// Menghitung selisih startDate dan endDate
-function dayDifference (start, end) {
-    const timeDiff = end.getTime() - start.getTime()
-    return Math.floor(timeDiff / (1000 * 60 * 60 * 24))
-}
-
-// Menentukan satuan waktu
-function chooseDuration(days) {
-    const years = Math.floor(days / 365)
-    const months = Math.floor((days % 365) / 30)
-    const remainingDays = days % 30
-
-    if(years > 0) {
-        return { duration: years, unit: 'tahun'}
-    } else if (months > 0) {
-        return { duration: months, unit: 'bulan'}
-    } else {
-        return { duration: remainingDays, unit: 'hari'}
-    }
-}
-
+// Start the server
 app.listen(port, () => {
     console.log(`Server berjalan pada port ${port}`)
 })
